src/app/guards/admin.guard.ts
import { Injectable } from "@angular/core";
import {
  CanActivate,
  ActivatedRouteSnapshot,
  RouterStateSnapshot,
  Router
} from "@angular/router";
import { Observable } from "rxjs";
import { User } from "../models/User";

@Injectable({
  providedIn: "root"
})
export class AdminGuard implements CanActivate {
  constructor(private router: Router) {}

  canActivate(): Observable<boolean> | Promise<boolean> | boolean {
    let user: User = JSON.parse(localStorage.getItem("user"));
    if (user && user.role.toLowerCase().includes("admin")) return true;
    else this.router.navigate([""]);
    return false;
  }
}
